My_Doctor/pyproject.toml
[tool.poetry]
name = "my-doctor"
version = "0.1.0"
description = "CRUD Aplication"
authors = ["VoitecP <piwowarski.connect@gmail.com>"]
readme = "../README.md"
packages = [{include = "my_doctor"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "*"
python-decouple = "*"
djangorestframework = "*"
markdown = "*"
django-filter = "*"
drf-spectacular = "*"
pycodestyle = "*"
pillow = "^10.1.0"
dj-rest-auth = "^5.0.1"
django-allauth = "^0.58.2"
djangorestframework-simplejwt = "^5.3.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
